Lab 9 - Spring Data

+++ Spring-lab-8/src/main/java/ua/lviv/iot/manegers/AirplaneRepository.java
@@ -7,6 +7,6 @@ import ua.lviv.iot.models.PassengerAircraft;
 //This will be AUTO IMPLEMENTED by Spring into a Bean called userRepository
 //CRUD refers Create, Read, Update, Delete
 
-public interface AirplaneRepository extends CrudRepository<PassengerAircraft, Integer> {
+public interface AirplaneRepository extends CrudRepository<PassengerAircraft, Long> {
 
 }
